+++ src/app/fire.rs
@@ -3,11 +3,11 @@ use app::timers::*;
 pub struct Fire {
     pub position: [u8; 2],
     pub fire_timer: Timer,
-    pub time_limit: u32,
+    pub time_limit: f32,
 }
 
 impl Fire {
-    pub fn create(position: [u8; 2], time_limit: u32) -> Self {
+    pub fn create(position: [u8; 2], time_limit: f32) -> Self {
         Fire {
             fire_timer: Timer::create(),
             position: position,
@@ -16,7 +16,7 @@ impl Fire {
     }
 
     pub fn ended(&self) -> bool {
-        self.fire_timer.did_pass((self.time_limit * 1000) as u64)
+        self.fire_timer.did_pass((self.time_limit * 1000.0) as u64)
     }
 
     pub fn get_position(&self) -> [u8; 2] {
